mrec_tools/pyproject.toml
[project]
name = "mrec_tools"
version = "1.0.0"
license = "GPL-3.0-or-later"
dependencies = [
    "pyyaml",
]
requires-python = ">= 3.11"
